feat: Add terraform to create test infra
This commit duplicates some of the tf setup I had in my demo env. This
is fine for now, I will modularise this later so we don't have almost
the same thing twice but my priority right now is getting this merged
and running in CI.

The `./scripts/tf.sh` has 3 options.
- `-v` will produce a vars.json for subsequent terraform commands to
  use. It will check for a metro with capacity as well as some other
  validation.
- `-u` (up) will run `terraform apply` with `--auto-approve`.
- `-d` (down) will run `terraform destroy` with `--auto-approve`.

The infra which is created on Equinix is almost the same as that which I
use for my demos. The main change is that there is no VPN set up. This
is because the VPN I used required a manual step (clicking a button in
Tailscale to approve subnet routes) which would not work here as we want
everything automated. The VPN was needed in my demo so that CAPMVM
clusters could be created from a local workstation without a) exposing
the infra publicly, and b) without having to reserve a public address to
serve as workload clusters' VIP LB address.

There are likely other, automatable ways to do this here, but I decided
not to get into that right now. So the tests which will eventually be
run against these Equinix hosts will be running within an Equinix
machine themselves for the immediate future.